libft v1

+++ ft_lstsize.c
@@ -6,15 +6,15 @@
 /*   By: amyburgh <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/03/27 23:58:03 by amyburgh          #+#    #+#             */
-/*   Updated: 2018/05/01 14:59:18 by amyburgh         ###   ########.fr       */
+/*   Updated: 2018/05/01 16:02:05 by amyburgh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_lstsize(t_list *begin_list)
+int	ft_lstsize(t_list *begin_list)
 {
-	size_t	count;
+	int	count;
 	t_list	*size;
 
 	count = 0;
